requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-filter==25.1
django-graphql-jwt==0.4.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
pillow==11.3.0
promise==2.3
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
